initial commit: setup failing test and spec
