+++ Java/gradle/wrapper/gradle-wrapper.properties
@@ -1,0 +1,6 @@
+#Wed Mar 14 19:28:13 CET 2018
+distributionBase=GRADLE_USER_HOME
+distributionPath=wrapper/dists
+zipStoreBase=GRADLE_USER_HOME
+zipStorePath=wrapper/dists
+distributionUrl=https\://services.gradle.org/distributions/gradle-3.5-rc-2-bin.zip
