+++ Java/src/main/java/com/evolvetechnology/TollCalculator.java
@@ -1,0 +1,21 @@
+package com.evolvetechnology;
+
+import com.evolvetechnology.vehicle.Vehicle;
+
+import java.time.LocalDateTime;
+
+public class TollCalculator {
+
+  /**
+   * Calculate the total toll fee for one day
+   *
+   * @param vehicle - the vehicle
+   * @param dates   - date and time of all passes on one day
+   * @return - the total toll fee for that day
+   */
+  public int getTollFee(Vehicle vehicle, LocalDateTime... dates) {
+    return 0;
+  }
+
+}
+
