+++ Java/src/main/java/com/evolvetechnology/vehicle/Car.java
@@ -1,0 +1,8 @@
+package com.evolvetechnology.vehicle;
+
+public class Car implements Vehicle {
+  @Override
+  public String getType() {
+    return "Car";
+  }
+}
