+++ Java/src/main/java/com/evolvetechnology/vehicle/Motorbike.java
@@ -1,0 +1,8 @@
+package com.evolvetechnology.vehicle;
+
+public class Motorbike implements Vehicle {
+  @Override
+  public String getType() {
+    return "Motorbike";
+  }
+}
