+++ Java/src/main/java/com/evolvetechnology/vehicle/Vehicle.java
@@ -1,0 +1,6 @@
+package com.evolvetechnology.vehicle;
+
+public interface Vehicle {
+
+  public String getType();
+}
